+++ nj/benches/nj_bench.rs
@@ -47,6 +47,8 @@ fn dist_config(msa: Vec<SequenceObject>) -> DistConfig {
         substitution_model: SubstitutionModel::JukesCantor,
         alphabet: None,
         num_threads: None,
+        gamma_shape: None,
+        p_invar: None,
     }
 }
 
@@ -59,6 +61,8 @@ fn nj_config(msa: Vec<SequenceObject>, n_bootstrap_samples: usize) -> NJConfig {
         num_threads: None,
         return_distance_matrix: false,
         return_average_distance: false,
+        gamma_shape: None,
+        p_invar: None,
     }
 }